AbstractBearerTransactionRequest
